requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Meta-learning
learn2learn==0.1.7

# Quantization
bitsandbytes>=0.41.0
accelerate>=0.24.0

# ML utilities
scikit-learn>=1.3.0
numpy>=1.24.0
scipy>=1.11.0

# Logging and visualization
tensorboard>=2.14.0
wandb>=0.15.0
tqdm>=4.66.0

# Configuration
pyyaml>=6.0
omegaconf>=2.3.0

# Data processing
pandas>=2.0.0
jsonlines>=4.0.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Code quality
black>=23.0.0
flake8>=6.1.0
isort>=5.12.0

# API (optional)
fastapi>=0.104.0
uvicorn>=0.24.0
